ME0 realistic digi updates
