+++ SimMuon/GEMDigitizer/interface/ME0PadDigiProducer.h
@@ -8,7 +8,7 @@
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Utilities/interface/InputTag.h"
 
-#include "DataFormats/GEMDigi/interface/ME0DigiPreRecoCollection.h"
+#include "DataFormats/GEMDigi/interface/ME0DigiCollection.h"
 #include "DataFormats/GEMDigi/interface/ME0PadDigiCollection.h"
 
 class ME0Geometry;
@@ -29,10 +29,10 @@ public:
 
 private:
 
-  void buildPads(const ME0DigiPreRecoCollection &digis, ME0PadDigiCollection &out_pads) const;
+  void buildPads(const ME0DigiCollection &digis, ME0PadDigiCollection &out_pads) const;
 
   /// Name of input digi Collection
-  edm::EDGetTokenT<ME0DigiPreRecoCollection> digi_token_;
+  edm::EDGetTokenT<ME0DigiCollection> digi_token_;
   edm::InputTag digis_;
 
   const ME0Geometry * geometry_;
