+++ SimMuon/GEMDigitizer/python/muonME0Digi_cff.py
@@ -1,6 +1,12 @@
 import FWCore.ParameterSet.Config as cms
 
+from SimMuon.GEMDigitizer.muonME0Digis_cfi import *
+from SimMuon.GEMDigitizer.muonME0PadDigis_cfi import *
 from SimMuon.GEMDigitizer.muonME0DigisPreReco_cfi import *
 from SimMuon.GEMDigitizer.muonME0ReDigis_cfi import *
 
-muonME0Digi = cms.Sequence(simMuonME0PseudoDigis*simMuonME0ReDigis)
+muonME0RealDigi = cms.Sequence(simMuonME0Digis * simMuonME0PadDigis)
+
+muonME0PseudoDigi = cms.Sequence(simMuonME0PseudoDigis * simMuonME0PseudoReDigis)
+
+muonME0Digi = cms.Sequence(muonME0RealDigi * muonME0PseudoDigi)
