+++ SimMuon/GEMDigitizer/python/muonME0PadDigis_cfi.py
@@ -2,5 +2,5 @@ import FWCore.ParameterSet.Config as cms
 
 # Module to create simulated ME0 trigger pad digis.
 simMuonME0PadDigis = cms.EDProducer("ME0PadDigiProducer",
-    InputCollection = cms.InputTag('simMuonME0ReDigis'),
+    InputCollection = cms.InputTag('simMuonME0Digis'),
 )
